+++ src/builtins/shared.rs
@@ -8,6 +8,7 @@ use crate::parser::{BlockType, LoopStatus};
 use crate::proc::{Pid, ProcStatus, no_exec};
 use crate::{builtins::*, wutil};
 use errno::errno;
+use fish_common::assert_sorted_by_name;
 use fish_wchar::L;
 use std::io::{BufRead, BufReader, Read};
